requirements.txt
APScheduler==3.11.0
asgiref==3.9.1
beautifulsoup4==4.14.2
Brotli==1.1.0
certifi==2025.7.14
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
charset-normalizer==3.4.2
cloudinary==1.44.1
cssselect2==0.8.0
daphne>=4.0,<5.0
dj-database-url==1.0.0
Django==5.2.4
django-cloudinary-storage==0.3.0
django-debug-toolbar==6.0.0
django-environ==0.12.0
django-htmx==1.23.2
django-redis>=6.0
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
fonttools==4.59.0
geopy==2.4.1
gunicorn==20.1.0
idna==3.10
openpyxl==3.1.5
pillow==11.3.0
psycopg[binary]>=3.1
pycparser==2.22
pydyf==0.11.0
pyphen==0.17.2
python-decouple==3.8
python-dotenv>=1.0.0
pytz==2025.2
pywebpush==2.0.3
redis==7.0.1
reportlab==4.4.3
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
urllib3==2.5.0
weasyprint==65.1
webencodings==0.5.1
whitenoise==6.5.0
zopfli==0.2.3.post1
